Update RADseq protocol with my recent adjustments
Remove the half library concentrations--we don't do half libraries and it creates chance for confusion. Lower digestion and PCR master mix volumes to 110 instead of 120 reactions--we don't need that many extra for loss, but leave the ligation MM at 120 reactions because the volumes are so small we DO need it. Add a 65C enzyme deactivation step to the ligation--it is in recent versions of the original protocol. Adjust PCR reaction for higher concentration of template--this seems to solve the high primer-dimer problem we were having.